webdisplay/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
pandas==2.1.3
scipy==1.11.4
python-socketio==5.10.0
python-dotenv==1.0.0